% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvcm.R
\name{gvcm}
\alias{gvcm}
\title{Generalized Varying Coefficient Model (GVCM) DML estimator (minimal paper version)}
\usage{
gvcm(
  data,
  Y,
  X,
  Z,
  link = c("gaussian", "binomial", "poisson"),
  K = 5,
  basis = c("ns", "poly", "none"),
  df = NULL,
  standardize_Z = TRUE,
  alpha = 0.5,
  nfolds_glmnet = 5,
  lambda_rule = c("lambda.min", "lambda.1se"),
  crossfit_seed = 1,
  eps_J = 1e-08,
  verbose = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{Y, X}{outcome/treatment: column name (character) or numeric vector}

\item{Z}{covariates: column names (character vector) or matrix/data.frame}

\item{link}{one of c("gaussian","binomial","poisson")}

\item{K}{number of cross-fitting folds}

\item{basis}{Basis specification for Z. Can be:
\itemize{
\item \code{"ns"}: natural spline basis applied column-wise,
\item \code{"poly"}: orthogonal polynomial basis applied column-wise,
\item \code{"none"}: use raw Z directly,
\item a formula, e.g. \code{~ Z1 + Z2 + Z1:Z2 + I(Z1^2)},
\item a function that takes the processed Z matrix and returns a numeric basis matrix,
\item or a precomputed numeric matrix/data.frame with the same number of rows as the analysis data.
}

Built-in options are convenient defaults. Custom specifications allow users
to include interactions, nonlinear transforms, or other user-defined basis terms
while leaving the rest of the estimation procedure unchanged.}

\item{df}{Degrees of freedom for the spline basi. Default is 4.}

\item{standardize_Z}{logical; center/scale raw Z before basis expansion}

\item{alpha}{Elastic-net mixing parameter used in the \pkg{glmnet} nuisance regressions.
Must be between 0 and 1. Defualt is 0.5.}

\item{nfolds_glmnet}{internal folds for cv.glmnet}

\item{lambda_rule}{c("lambda.min","lambda.1se")}

\item{crossfit_seed}{seed for cross-fitting folds}

\item{eps_J}{lower truncation for \eqn{\hat J(Z)} to avoid division by zero}

\item{verbose}{logical}
}
\value{
A list with \code{theta_hat}, \code{se_hat}, and \code{ci}.
}
\description{
Fits a generalized varying coefficient model under a canonical GLM link:
\eqn{g(\mathbb{E}[Y \mid X, Z]) = \beta_0(Z) + \beta_1(Z)\,X}
and estimates the target parameter
\eqn{\theta = \mathbb{E}[\beta_1(Z)]}
using cross-fitted Double Machine Learning (DML).
}
